requirements.txt
streamlit==1.22.0
pillow==9.3.0
numpy==1.21.6
requests==2.28.1
